no errors, css needs work

+++ app/models/product.py
@@ -19,7 +19,7 @@ class Product(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer, db.ForeignKey(add_prefix_for_prod("users.id")), nullable=False)
     title = db.Column(db.String(40), unique=True, nullable = False)
-    price = db.Column(db.Integer(), nullable=False)
+    price = db.Column(db.Float(), nullable=False)
     description = db.Column(db.String(1000), nullable = False)
     glitter_factor = db.Column(db.String(600), nullable = False)
     product_image = db.Column(db.String(300), nullable = True)
